+++ src/LighthouseServiceProvider.php
@@ -33,6 +33,7 @@ use Nuwave\Lighthouse\Execution\LighthouseRequest;
 use Nuwave\Lighthouse\Execution\MultipartFormRequest;
 use Nuwave\Lighthouse\Execution\SingleResponse;
 use Nuwave\Lighthouse\Execution\Utils\GlobalId;
+use Nuwave\Lighthouse\Execution\ValidationRulesProvider;
 use Nuwave\Lighthouse\Schema\AST\ASTBuilder;
 use Nuwave\Lighthouse\Schema\Factories\DirectiveFactory;
 use Nuwave\Lighthouse\Schema\NodeRegistry;
@@ -51,6 +52,7 @@ use Nuwave\Lighthouse\Support\Contracts\CreatesResponse;
 use Nuwave\Lighthouse\Support\Contracts\GlobalId as GlobalIdContract;
 use Nuwave\Lighthouse\Support\Contracts\ProvidesResolver;
 use Nuwave\Lighthouse\Support\Contracts\ProvidesSubscriptionResolver;
+use Nuwave\Lighthouse\Support\Contracts\ProvidesValidationRules;
 use Nuwave\Lighthouse\Support\Http\Responses\ResponseStream;
 use Nuwave\Lighthouse\Testing\TestingServiceProvider;
 
@@ -63,11 +65,11 @@ class LighthouseServiceProvider extends ServiceProvider
     {
         $this->publishes([
             __DIR__.'/lighthouse.php' => $this->app->configPath().'/lighthouse.php',
-        ], 'config');
+        ], 'lighthouse-config');
 
         $this->publishes([
             __DIR__.'/../assets/default-schema.graphql' => $configRepository->get('lighthouse.schema.register'),
-        ], 'schema');
+        ], 'lighthouse-schema');
 
         $this->loadRoutesFrom(__DIR__.'/Support/Http/routes.php');
 
@@ -145,6 +147,8 @@ class LighthouseServiceProvider extends ServiceProvider
                 }
             };
         });
+
+        $this->app->bind(ProvidesValidationRules::class, ValidationRulesProvider::class);
 
         $this->app->singleton(MiddlewareAdapter::class, function (Container $app): MiddlewareAdapter {
             // prefer using fully-qualified class names here when referring to Laravel-only or Lumen-only classes